feat/fix: add ode history and fix solver errors
- create handle class for storing previous time step solutions when
solving the ode. The convergeEvent function use the past steps to more
robustly detect convergence in the case of oscillating particles in
the clusters.
- add convergeEvent in the case of NaN values.
- if NaN values are detected the simulation will be restart from a
previous solution with a new initial time step of half the size. This
fixes cases where the very initial time step was too large, and cases
where the time step increases near the end of the simulation, but then
a particle jumps out from the simulation domain.