Feat: Add grid_to_data to problem_scales
Problem: grid_spacing needs to be modified for making the confining
potential scale invariant for nuclei. grid_spacing was also
used to convert between grid and data spaces; however, the actual
scale factor converting between grid and data spaces should not be
scaled.
Correction: add new field to problem_scales, grid_to_data, that gives
the scale factor to convert from grid space to data space -- these
factors are not scaled when scaling grid_spacing. This is the factor
that should be used when converting points between grid and data spaces.
Result: The performance of the method increased back to what is was
before - ~0.915 F1_3 and ~0.915 dN0.