Feat: Add functions for converting between grid and data
Problem: The previous conversions between grid and data did not include
the translation of origin between the two spaces, but only the scale
factors.
Correction: computeProblemScales() should now recieve the data limits
and not the data range. The problem_scales structure now has
data_to_grid and grid_to_data fields that hold function handles to
convert between the two spaces, and take the Potential_Padding_Size into
account.

Also changed name of grid_range to grid_size. The size of the grid
should be given, not the size minus 1.