Feat: Add verbose property and class to handle output
Problem: checking if we should output information on iterations and then
computing, formatting, and displaying that information takes many lines
of code and detracts from ease of viewing.
Correction:
* Added property Verbose to seedPointOptions()
* Created new class displayProgress() that handles computing the
iteration number, the time the loops have been running, and displaying
messages every so often. If the matlab version is 9.2 or above, it also
allows for displaying messages in parfor loops.