Feat: Add ability to have empty options to computeProblemScales
Problem: When computing the centroid in computeObjectSeedPoints, the
resulting centroid is in grid units, but the grid has not been padded
with Potential_Padding_Size. We need the grid->data conversion with a
padding size of zero.
Correction: Add default pad size of 0 when options is empty. Also, set
the attractive extents to 1, when options is empty.